requirements.txt
kafka-python
faker
pandas
openpyxl
psycopg2-binary 

